Valve files are moved to Hradware/Valves
