+++ Irrigation/include/Logic/WateringController.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include "RunTime.h"
 #include "RunTimeContainer.h"
-#include "Hardware/ZoneHandler.h"
+#include "Hardware/Valves/ZoneHandler.h"
 
 
 class WateringController {
