+++ Irrigation/lib/Logic/WateringController.cpp
@@ -1,7 +1,7 @@
 #include "WateringController.h"
 #include <stdexcept>
 #include <thread>
-#include "Hardware/ZoneHandler.h"
+#include "Hardware/Valves/ZoneHandler.h"
 #include "Logger/Logger.h"
 #include "Utils/ToString.h"
