+++ Irrigation/test/Hardware/Valves/ZoneHandlerTest.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <gmock/gmock.h>
 #include <memory>
-#include "Hardware/Valve.h"
-#include "Hardware/ZoneHandler.h"
+#include "Hardware/Valves/Valve.h"
+#include "Hardware/Valves/ZoneHandler.h"
 
 
 class MockValve : public Valve {
