+++ Irrigation/test/Model/IrrigationDocumentTest.cpp
@@ -4,8 +4,8 @@
 #include "IrrigationDocumentTest.h"
 #include "DTO/DocumentDTO.h"
 #include "Exceptions/Exceptions.h"
-#include "Hardware/GpioHandler.h"
-#include "Hardware/ZoneHandler.h"
+#include "Hardware/Valves/GpioHandler.h"
+#include "Hardware/Valves/ZoneHandler.h"
 #include "Logic/Program.h"
 #include "Logic/RunTime.h"
 #include "Logic/StartTime.h"
