+++ Irrigation/test/Views/TimerView/Makefile
@@ -20,7 +20,7 @@ DEPENDS         = $(patsubst $(SOURCE_DIR)/%,$(OBJECT_DIR)/%,$(SOURCES:.$(SRC_EX
 ########################################################
 ########################################################
 
-LOCAL_LIBS      = -l$(TARGET_NAME) -lLogger -lModel -lDtoReaderWriter -lDocumentView -lLogic -lDTO -lUtils -lSchedulers -lHardware
+LOCAL_LIBS      = -l$(TARGET_NAME) -lLogger -lModel -lDtoReaderWriter -lDocumentView -lLogic -lDTO -lUtils -lSchedulers -lValves
 INCLUDE_PATH    = -I$(INCLUDE_DIR)
 LIBDIRS         = -L$(LIBRARY_DIR)
 LIBS            = $(LOCAL_LIBS) -pthread -lgmock -lpugixml
